made conversion to OOP
