+++ src/components/filter.js
@@ -1,59 +1,30 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const getTripEventsFilterForm = () => {
   return (
-    `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-        <span class="trip-sort__item  trip-sort__item--day">Day</span>
+    `<form class="trip-filters" action="#" method="get">
+            <div class="trip-filters__filter">
+            <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
+            <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
+            </div>
 
-        <div class="trip-sort__item  trip-sort__item--event">
-        <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" checked>
-        <label class="trip-sort__btn" for="sort-event">Event</label>
-        </div>
+            <div class="trip-filters__filter">
+            <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
+            <label class="trip-filters__filter-label" for="filter-future">Future</label>
+            </div>
 
-        <div class="trip-sort__item  trip-sort__item--time">
-        <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
-        <label class="trip-sort__btn" for="sort-time">
-            Time
-            <svg class="trip-sort__direction-icon" width="8" height="10" viewBox="0 0 8 10">
-            <path d="M2.888 4.852V9.694H5.588V4.852L7.91 5.068L4.238 0.00999987L0.548 5.068L2.888 4.852Z"/>
-            </svg>
-        </label>
-        </div>
+            <div class="trip-filters__filter">
+            <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
+            <label class="trip-filters__filter-label" for="filter-past">Past</label>
+            </div>
 
-        <div class="trip-sort__item  trip-sort__item--price">
-        <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
-        <label class="trip-sort__btn" for="sort-price">
-            Price
-            <svg class="trip-sort__direction-icon" width="8" height="10" viewBox="0 0 8 10">
-            <path d="M2.888 4.852V9.694H5.588V4.852L7.91 5.068L4.238 0.00999987L0.548 5.068L2.888 4.852Z"/>
-            </svg>
-        </label>
-        </div>
-
-        <span class="trip-sort__item  trip-sort__item--offers">Offers</span>
-    </form>`
+            <button class="visually-hidden" type="submit">Accept filter</button>
+        </form>`
   );
 };
 
-
-export default class Filter {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Filter extends AbstractComponent {
   getTemplate() {
     return getTripEventsFilterForm();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
