+++ src/components/no-trips.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const getNoTripMessage = function () {
   return (
@@ -6,25 +6,8 @@ const getNoTripMessage = function () {
   );
 };
 
-
-export default class Message {
-  constructor() {
-    this._element = null;
-  }
-
+export default class Message extends AbstractComponent {
   getTemplate() {
     return getNoTripMessage();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
