+++ src/components/trip-add.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const getOffersMarkup = (offers) => {
   return offers
@@ -166,24 +166,24 @@ const getTripEventsAdd = (trip) => {
   );
 };
 
-export default class TripAdd {
+export default class TripAdd extends AbstractComponent {
   constructor(trip) {
+    super();
+
     this._trip = trip;
-    this._element = null;
   }
 
   getTemplate() {
     return getTripEventsAdd(this._trip);
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
+  setCloseButtonClickHandler(handler) {
+    this.getElement().querySelector(`.event__rollup-btn`)
+      .addEventListener(`click`, handler);
   }
 
-  removeElement() {
-    this._element = null;
+  formSubmit(handler) {
+    this.getElement().querySelector(`form`)
+      .addEventListener(`submit`, handler);
   }
 }
