+++ src/components/trips.js
