+++ src/utils/render.js
@@ -1,40 +1,39 @@
 export const RenderPosition = {
-    AFTERBEGIN: `afterbegin`,
-    BEFOREEND: `beforeend`
-  };
-  
-  export const createElement = (template) => {
-    const newElement = document.createElement(`div`);
-    newElement.innerHTML = template;
-  
-    return newElement.firstChild;
-  };
-  
-  export const render = (container, component, place) => {
-    switch (place) {
-      case RenderPosition.AFTERBEGIN:
-        container.prepend(component.getElement());
-        break;
-      case RenderPosition.BEFOREEND:
-        container.append(component.getElement());
-        break;
-    }
-  };
-  
-  export const remove = (component) => {
-    component.getElement().remove();
-    component.removeElement();
-  };
-  
-  export const replace = (newComponent, oldComponent) => {
-    const parentElement = oldComponent.getElement().parentElement;
-    const newElement = newComponent.getElement();
-    const oldElement = oldComponent.getElement();
-  
-    const isExistElements = !!(parentElement && newElement && oldElement);
-  
-    if (isExistElements && parentElement.contains(oldElement)) {
-      parentElement.replaceChild(newElement, oldElement);
-    }
-  };
-  
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+export const render = (container, component, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(component.getElement());
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(component.getElement());
+      break;
+  }
+};
+
+export const remove = (component) => {
+  component.getElement().remove();
+  component.removeElement();
+};
+
+export const replace = (newComponent, oldComponent) => {
+  const parentElement = oldComponent.getElement().parentElement;
+  const newElement = newComponent.getElement();
+  const oldElement = oldComponent.getElement();
+
+  const isExistElements = !!(parentElement && newElement && oldElement);
+
+  if (isExistElements && parentElement.contains(oldElement)) {
+    parentElement.replaceChild(newElement, oldElement);
+  }
+};
